+++ src/main/java/racingcar/controller/RacingCarController.java
@@ -1,11 +1,13 @@
 package racingcar.controller;
 
 import racingcar.service.CarService;
+import racingcar.service.MoveCountService;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class RacingCarController {
     private final CarService service = new CarService();
+    private final MoveCountService moveCountService = new MoveCountService();
 
     public void handleInputCarNames() {
         InputView.printCarNamesInputGuide();
@@ -14,7 +16,7 @@ public class RacingCarController {
 
     public void handleInputMoveCount() {
         InputView.printMoveCountInputGuide();
-        service.setMoveCount(InputView.inputMoveCount());
+        moveCountService.setMoveCount(InputView.inputMoveCount());
     }
 
     public void startGame() {
@@ -22,7 +24,7 @@ public class RacingCarController {
         handleInputMoveCount();
 
         OutputView.printRunMessage();
-        for (int i = 0; i < service.getMoveCount(); i++) {
+        for (int i = 0; i < moveCountService.getMoveCount(); i++) {
             service.playGame();
             OutputView.printGameStatus(service.getCarList().getCars());
         }
