+++ src/test/java/racingcar/service/MoveCountTest.java
@@ -11,10 +11,10 @@ public class MoveCountTest {
     @DisplayName("시도할 횟수가 1 미만일 때 예외가 발생해야 한다")
     void setMoveCount_lessThanOne_shouldThrowException() {
         // given
-        CarService carService = new CarService();
+        MoveCountService moveCountService = new MoveCountService();
 
         // when & then
-        assertThatThrownBy(() -> carService.setMoveCount(String.valueOf(0)))
+        assertThatThrownBy(() -> moveCountService.setMoveCount(String.valueOf(0)))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(Constants.INVALID_MOVE_COUNT);
     }
@@ -23,11 +23,11 @@ public class MoveCountTest {
     @DisplayName("시도 횟수가 숫자가 아닐 때 예외가 발생해야 한다")
     void setMoveCount_whenNotANumber_shouldThrowException() {
         // Given
-        CarService carService = new CarService();
+        MoveCountService moveCountService = new MoveCountService();
         String invalidMoveCount = "three";  // 숫자가 아닌 문자열 입력
 
         // When & Then
-        assertThatThrownBy(() -> carService.setMoveCount(invalidMoveCount))
+        assertThatThrownBy(() -> moveCountService.setMoveCount(invalidMoveCount))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage(Constants.INVALID_MOVE_COUNT_NUMBER);
     }
